requirements.txt
arxiv==2.1.3
google-ai-generativelanguage==0.6.10
google-api-core==2.20.0
google-api-python-client==2.147.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.2
googleapis-common-protos==1.65.0
MarkupSafe==2.1.5
ollama==0.3.3
PyMuPDF==1.24.10
PyMuPDFb==1.24.10
pypdf==5.0.1
PyPDF2==3.0.1
requests==2.32.3
